requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.7.22
click==8.1.7
fastapi==0.104.1
grpcio==1.59.3
grpcio-testing==1.59.3
grpcio-tools==1.59.3
grpclib==0.4.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.1
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
multidict==6.0.4
mypy==1.6.1
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pypika-tortoise==0.1.6
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
bcrypt==4.1.1
